Settings: fix profiles TabManager
Whenever an user enters in the profiles settings a new TabHost, but the TabManager is neither created
or updated with this new tabhost. On tablets when the user returns to profiles from a fragment
activity, the tabhost is not displayed.
TabManager need to be created along with the tabhost.

Patchset 2: Fixed ProfileEnabler on tablets. The instance of ProfileEnabler of Settings that
            is the one used on tablets doesn't have a reference of ProfileSettings, so this class
            never will be updated. With this pathset the profiles enabler follows the rules of
            wifi and bluetooth enablers: broadcast intent. Now, settings listen for
            PROFILES_STATE_CHANGED_ACTION events, and refresh its own ui.

Requires: http://review.cyanogenmod.org/#/c/32460/1

Change-Id: I5eae79cd20cee3c14f7d2c30a2238425ccd15e09
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>