Lets save graphite mode for a time when its more applicable...
This makes the screen dark, black helps amoleds, for now lets just name it dark mode
Graphite mode ON would almost make sense as the default as the default Holo theme is grey...
kinda like graphite. Anyways i think this is a more apt description. I understand there was some difficulty
getting this to integrate into a deeper level of the settings but we should get on that soon...
WARNING THIS WAS ALTERED ON A CLONE AND WAS NOT TESTED (broke my own rule) Im about to run it through the compiler
and assuming all is well it will remain, if not a quick fix is on the way.