Hidden Statusbar Pulldown (2/2)
Settings part for making "Swiping on the top of full screen apps shows status bar for 10 seconds"
an on/off option and make Auto Hide Statusbar a dependency. We also need to be more consistent
with our titles. Feel free to revert that part if yall dont agree :)

Big thanks to blk_jack for helping me on this!