Statusbar Hide and Seek (2/2)
Credit Bajee11 (RootBox)

Statusbar Behaviour: Switch to list preference to avoid confusion (2/2)

Conflicts:
	res/values/arrays.xml

Change-Id: I49bd1f119ed0a21573ba46ed27cd4e6f35e34123

Rework StatusBar logic (2/2)

Change-Id: I70b0337e17f3c47317003c7dec7bf94ee161e019

RC: Hidden Statusbar Pulldown (2/2)

Currently this gives the options to turn it on/off and configure the timeout
for the Statusbar before hiding again. This is dependent on Hide Statusbar
being on since it doesnt make sense to have on without it. Thats also why I
made the changes to the mStatusBarHide part so it turns off when Hide Statusbar
is unchecked. If its written badly or whatever, please change however needed
since I suck at java.

Change-Id: I9ddd2d1a41da42eb264494872c5928aad8397933